requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.50.0
